hr_system_enterprise_v2/requirements.txt
streamlit
pandas
numpy
google-generativeai
python-docx
xlsxwriter
pydantic
fastapi
uvicorn